feat(event-display): create a url options manager for url options
